Optionally support retrieving AWS credentials from standard locations.
There are several standard locations for AWS credentials, like environment
variables and dotfiles. Previously this plugin ignored those and expected the
user to specify the access & secret key.

Update dynamodb-client.js to use the values set in config, if available, and if
not default to not setting them. Not setting the credentials when creating the
AWS.DynamoDB instance causes the AWS sdk to look for the credentials in the
standard locations, like environment variables, dotfile, and even ec2 instance
profiles.

http://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/setting-credentials-node.html